Move HTML generator into its own module
